+++ client/src/apis/request/report.ts
@@ -4,7 +4,7 @@ import {WithErrorHandlingStrategy} from '@errors/RequestGetError';
 
 import {BASE_URL} from '@apis/baseUrl';
 import {MEMBER_API_PREFIX} from '@apis/endpointPrefix';
-import {requestGet} from '@apis/fetcher';
+import {requestGet} from '@apis/request';
 import {WithEventId} from '@apis/withId.type';
 
 export const requestGetReports = async ({eventId, ...props}: WithEventId<WithErrorHandlingStrategy>) => {
