+++ client/src/apis/request/step.ts
@@ -3,7 +3,7 @@ import {WithErrorHandlingStrategy} from '@errors/RequestGetError';
 
 import {BASE_URL} from '@apis/baseUrl';
 import {MEMBER_API_PREFIX} from '@apis/endpointPrefix';
-import {requestGet} from '@apis/fetcher';
+import {requestGet} from '@apis/request';
 import {WithEventId} from '@apis/withId.type';
 
 export const requestGetSteps = async ({eventId, ...props}: WithEventId<WithErrorHandlingStrategy>) => {
